Move pkg/api/v1/ref -> client-go/tools/reference
