+++ pkg/controller/node/testutil/test_utils.go
@@ -34,7 +34,7 @@ import (
 
 	clientv1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/util/clock"
-	"k8s.io/kubernetes/pkg/api/v1/ref"
+	ref "k8s.io/client-go/tools/reference"
 
 	"k8s.io/api/core/v1"
 	"k8s.io/client-go/kubernetes/fake"
