+++ staging/src/k8s.io/client-go/tools/reference/ref.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package ref
+package reference
 
 import (
 	"errors"
